Move: TermsTemplate 폴더 이동 및 불필요 코드 제거
